+++ packages/terra-framework-docs/src/terra-dev-site/test/compact-interactive-list/InconsistentWidthUnits.test.jsx
@@ -75,7 +75,7 @@ const rows = [
 const cols = [
   {
     id: 'Column-0',
-    displayName: 'Icon',
+    displayName: 'Status',
     width: '60px',
     align: alignTypes.CENTER,
     maximumWidth: '10em',