+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Updated `terra-compact-interactive-list` examples to provide a meaningful name to the column with icons.
+
 ## 1.78.0 - (March 29, 2024)
 
 * Changed